+++ tests/Utility/SkipIntegrationException.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Copyright (c) Precision Soft
+ */
+
+namespace PrecisionSoft\Doctrine\Type\Test\Utility;
+
+use RuntimeException;
+
+/** @internal */
+final class SkipIntegrationException extends RuntimeException {}